+++ src/mcp/resources.ts
@@ -10,8 +10,8 @@ import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { ResourceTemplate } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { ErrorCode, McpError } from "@modelcontextprotocol/sdk/types.js";
 import { remoteClient } from "../activitypub/remote-client.js";
+import type { RateLimiter } from "../resilience/rate-limiter.js";
 import { extractSingleValue, validateActorIdentifier } from "../server/index.js";
-import type { RateLimiter } from "../server/rate-limiter.js";
 import { getErrorMessage } from "../utils.js";
 import { DomainSchema } from "../validation/schemas.js";
 import { capabilitiesRegistry, trackedMcpServer } from "./capabilities.js";
