+++ src/mcp/tools-export.ts
@@ -10,8 +10,8 @@ import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { ErrorCode, McpError } from "@modelcontextprotocol/sdk/types.js";
 import { z } from "zod";
 import { remoteClient } from "../activitypub/remote-client.js";
+import type { RateLimiter } from "../resilience/rate-limiter.js";
 import { validateActorIdentifier, validateDomain } from "../server/index.js";
-import type { RateLimiter } from "../server/rate-limiter.js";
 import { performanceMonitor } from "../telemetry/performance-monitor.js";
 import { formatErrorWithSuggestion, getErrorMessage, stripHtmlTags } from "../utils.js";
 import { trackedMcpServer } from "./capabilities.js";
