+++ src/server/index.ts
@@ -6,9 +6,9 @@ export {
   AdaptiveRateLimiter,
   adaptiveRateLimiter,
   type InstanceRateLimit,
-} from "./adaptive-rate-limiter.js";
+} from "../resilience/adaptive-rate-limiter.js";
+export { type RateLimitConfig, RateLimiter } from "../resilience/rate-limiter.js";
 export { type HttpTransportOptions, HttpTransportServer } from "./http-transport.js";
-export { type RateLimitConfig, RateLimiter } from "./rate-limiter.js";
 export {
   extractSingleValue,
   validateActorIdentifier,
